refactor(ui): Extract runs table to separate component
