+++ ui/src/routes/_layout/organizations/$orgId/products/$productId/repositories/$repoId/index.tsx
@@ -18,31 +18,20 @@
  */
 
 import { createFileRoute, Link } from '@tanstack/react-router';
-import {
-  createColumnHelper,
-  getCoreRowModel,
-  useReactTable,
-} from '@tanstack/react-table';
-import { EditIcon, PlusIcon, Repeat, View } from 'lucide-react';
+import { EditIcon, PlusIcon } from 'lucide-react';
 
 import {
   useRepositoriesServiceDeleteRepositoryById,
-  useRepositoriesServiceGetOrtRunsByRepositoryId,
   useRepositoriesServiceGetRepositoryById,
 } from '@/api/queries';
 import {
   prefetchUseRepositoriesServiceGetOrtRunsByRepositoryId,
   prefetchUseRepositoriesServiceGetRepositoryById,
 } from '@/api/queries/prefetch';
-import { ApiError, OrtRunSummary } from '@/api/requests';
-import { DataTable } from '@/components/data-table/data-table';
+import { ApiError } from '@/api/requests';
 import { DeleteDialog } from '@/components/delete-dialog';
 import { LoadingIndicator } from '@/components/loading-indicator';
-import { OrtRunJobStatus } from '@/components/ort-run-job-status';
-import { RunDuration } from '@/components/run-duration';
-import { TimestampWithUTC } from '@/components/timestamp-with-utc';
 import { ToastError } from '@/components/toast-error';
-import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
 import {
   Card,
@@ -56,128 +45,11 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from '@/components/ui/tooltip';
-import { config } from '@/config';
-import { getStatusBackgroundColor } from '@/helpers/get-status-class';
 import { toast } from '@/lib/toast';
 import { paginationSearchParameterSchema } from '@/schemas';
+import { RepositoryRunsTable } from './-components/repository-runs-table';
 
 const defaultPageSize = 10;
-const pollInterval = config.pollInterval;
-
-const columnHelper = createColumnHelper<OrtRunSummary>();
-
-const columns = [
-  columnHelper.accessor('index', {
-    header: 'Run',
-    cell: ({ row }) => (
-      <Link
-        className='font-semibold text-blue-400 hover:underline'
-        to={
-          '/organizations/$orgId/products/$productId/repositories/$repoId/runs/$runIndex'
-        }
-        params={{
-          orgId: row.original.organizationId.toString(),
-          productId: row.original.productId.toString(),
-          repoId: row.original.repositoryId.toString(),
-          runIndex: row.original.index.toString(),
-        }}
-      >
-        {row.original.index}
-      </Link>
-    ),
-    size: 50,
-  }),
-  columnHelper.accessor('createdAt', {
-    header: 'Created At',
-    cell: ({ row }) => <TimestampWithUTC timestamp={row.original.createdAt} />,
-  }),
-  columnHelper.accessor('status', {
-    header: 'Run Status',
-    cell: ({ row }) => (
-      <Badge
-        className={`border ${getStatusBackgroundColor(row.original.status)}`}
-      >
-        {row.original.status}
-      </Badge>
-    ),
-  }),
-  columnHelper.display({
-    id: 'jobStatuses',
-    header: () => <div>Job Status</div>,
-    cell: ({ row }) => (
-      <OrtRunJobStatus
-        jobs={row.original.jobs}
-        orgId={row.original.organizationId.toString()}
-        productId={row.original.productId.toString()}
-        repoId={row.original.repositoryId.toString()}
-        runIndex={row.original.index.toString()}
-      />
-    ),
-  }),
-  // TODO: Write this with an accessor as soon as I know how to do it.
-  columnHelper.display({
-    id: 'duration',
-    header: 'Duration',
-    cell: ({ row }) => (
-      <RunDuration
-        createdAt={row.original.createdAt}
-        finishedAt={row.original.finishedAt ?? undefined}
-      />
-    ),
-  }),
-  columnHelper.display({
-    id: 'actions',
-    header: () => <div>Actions</div>,
-    cell: ({ row }) => (
-      <div className='flex gap-2'>
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <Button variant='outline' asChild size='sm'>
-              <Link
-                to={
-                  '/organizations/$orgId/products/$productId/repositories/$repoId/runs/$runIndex'
-                }
-                params={{
-                  orgId: row.original.organizationId.toString(),
-                  productId: row.original.productId.toString(),
-                  repoId: row.original.repositoryId.toString(),
-                  runIndex: row.original.index.toString(),
-                }}
-              >
-                View
-                <View className='ml-1 h-4 w-4' />
-              </Link>
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent>View the details of this run</TooltipContent>
-        </Tooltip>
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <Button variant='outline' asChild size='sm'>
-              <Link
-                to='/organizations/$orgId/products/$productId/repositories/$repoId/create-run'
-                params={{
-                  orgId: row.original.organizationId.toString(),
-                  productId: row.original.productId.toString(),
-                  repoId: row.original.repositoryId.toString(),
-                }}
-                search={{
-                  rerunIndex: row.original.index,
-                }}
-              >
-                Rerun
-                <Repeat className='ml-1 h-4 w-4' />
-              </Link>
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent>
-            Create a new ORT run based on this run
-          </TooltipContent>
-        </Tooltip>
-      </div>
-    ),
-  }),
-];
 
 const RepoComponent = () => {
   const params = Route.useParams();
@@ -194,24 +66,6 @@ const RepoComponent = () => {
   } = useRepositoriesServiceGetRepositoryById({
     repositoryId: Number.parseInt(params.repoId),
   });
-
-  const {
-    data: runs,
-    error: runsError,
-    isPending: runsIsPending,
-    isError: runsIsError,
-  } = useRepositoriesServiceGetOrtRunsByRepositoryId(
-    {
-      repositoryId: Number.parseInt(params.repoId),
-      limit: pageSize,
-      offset: pageIndex * pageSize,
-      sort: '-index',
-    },
-    undefined,
-    {
-      refetchInterval: pollInterval,
-    }
-  );
 
   const { mutateAsync: deleteRepository, isPending } =
     useRepositoriesServiceDeleteRepositoryById({
@@ -242,27 +96,13 @@ const RepoComponent = () => {
     });
   }
 
-  const table = useReactTable({
-    data: runs?.data || [],
-    columns,
-    pageCount: Math.ceil((runs?.pagination.totalCount ?? 0) / pageSize),
-    state: {
-      pagination: {
-        pageIndex,
-        pageSize,
-      },
-    },
-    getCoreRowModel: getCoreRowModel(),
-    manualPagination: true,
-  });
-
-  if (repoIsPending || runsIsPending) {
+  if (repoIsPending) {
     return <LoadingIndicator />;
   }
 
-  if (repoIsError || runsIsError) {
+  if (repoIsError) {
     toast.error('Unable to load data', {
-      description: <ToastError error={repoError || runsError} />,
+      description: <ToastError error={repoError} />,
       duration: Infinity,
       cancel: {
         label: 'Dismiss',
@@ -335,20 +175,11 @@ const RepoComponent = () => {
         </div>
       </CardHeader>
       <CardContent>
-        <DataTable
-          table={table}
-          setCurrentPageOptions={(currentPage) => {
-            return {
-              to: Route.to,
-              search: { ...search, page: currentPage },
-            };
-          }}
-          setPageSizeOptions={(size) => {
-            return {
-              to: Route.to,
-              search: { ...search, page: 1, pageSize: size },
-            };
-          }}
+        <RepositoryRunsTable
+          repoId={params.repoId}
+          pageIndex={pageIndex}
+          pageSize={pageSize}
+          search={search}
         />
       </CardContent>
     </Card>
